requirements.txt
discord
python-dotenv
yt-dlp